# api/swagger.json
{
	"openapi": "3.0.0",
	"paths": {
		"/users/{userName}": {
			"get": {
				"operationId": "UsersController_getUserByUserName",
				"parameters": [
					{
						"name": "userName",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": true
												},
												"data": {
													"$ref": "#/components/schemas/GetUserDto"
												}
											},
											"required": [
												"success",
												"data"
											]
										},
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": false
												},
												"data": {
													"type": "string"
												}
											},
											"required": [
												"success",
												"data"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				]
			}
		},
		"/users/me": {
			"post": {
				"operationId": "UsersController_getMe",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MeDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": true
												},
												"data": {
													"$ref": "#/components/schemas/GetMeDto"
												}
											},
											"required": [
												"success",
												"data"
											]
										},
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": false
												},
												"data": {
													"type": "string"
												}
											},
											"required": [
												"success",
												"data"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				]
			}
		},
		"/users/create": {
			"post": {
				"operationId": "UsersController_createUser",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": true
												},
												"data": {
													"$ref": "#/components/schemas/UserEntity"
												}
											},
											"required": [
												"success",
												"data"
											]
										},
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": false
												},
												"data": {
													"type": "string"
												}
											},
											"required": [
												"success",
												"data"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				]
			}
		},
		"/users/update": {
			"patch": {
				"operationId": "UsersController_updateUser",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": true
												},
												"data": {
													"type": "object"
												}
											},
											"required": [
												"success",
												"data"
											]
										},
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": false
												},
												"data": {
													"type": "string"
												}
											},
											"required": [
												"success",
												"data"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				]
			}
		},
...
	},
	"info": {
		"title": "PassimX",
		"description": "",
		"version": "1.1.3",
		"contact": {},
		"license": {
			"name": "MIT LICENSE",
			"url": "open platform for messaging"
		}
	},
...
	"servers": [
		{
			"url": "http://localhost:7020"
		}
	],
	"components": {
		"schemas": {
			"GetUserDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"rsaPublicKey": {
						"type": "string"
					},
					"createdAt": {
						"format": "date-time",
						"type": "string"
					}
				},
				"required": [
					"name",
					"userName",
					"rsaPublicKey",
					"createdAt"
				]
			},
			"GetMeDto": {
				"type": "object",
				"properties": {
					"encryptedRsaPrivateKey": {
						"type": "string"
					}
				},
				"required": [
					"encryptedRsaPrivateKey"
				]
			},
			"MeDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"seedPhraseHash": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"seedPhraseHash"
				]
			},
			"UserEntity": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"rsaPublicKey": {
						"type": "string"
					},
					"encryptedRsaPrivateKey": {
						"type": "string"
					},
					"createdAt": {
						"format": "date-time",
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"userName",
					"rsaPublicKey",
					"encryptedRsaPrivateKey",
					"createdAt"
				]
			},
			"CreateUserDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"rsaPublicKey": {
						"type": "string"
					},
					"encryptedRsaPrivateKey": {
						"type": "string"
					},
					"passwordHash": {
						"type": "string"
					},
					"seedPhraseHash": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"rsaPublicKey",
					"encryptedRsaPrivateKey",
					"passwordHash",
					"seedPhraseHash"
				]
			},
			"UpdateDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"seedPhraseHash": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"seedPhraseHash",
					"name"
				]
			},
...
		}
	}
}